+++ tests/integration/modules/test_cmdmod.py
@@ -5,7 +5,6 @@ from __future__ import absolute_import
 import os
 import sys
 import textwrap
-import tempfile
 
 # Import Salt Testing libs
 import tests.integration as integration
@@ -14,7 +13,6 @@ from tests.support.helpers import (
     destructiveTest,
     skip_if_binaries_missing
 )
-from tests.support.mock import NO_MOCK, NO_MOCK_REASON, Mock, patch
 
 # Import salt libs
 import salt.utils
@@ -32,7 +30,6 @@ AVAILABLE_PYTHON_EXECUTABLE = salt.utils.which_bin([
 ])
 
 
-@skipIf(NO_MOCK, NO_MOCK_REASON)
 class CMDModuleTest(integration.ModuleCase):
     '''
     Validate the cmd module
@@ -65,45 +62,6 @@ class CMDModuleTest(integration.ModuleCase):
         self.assertEqual(self.run_function('cmd.run',
                          ['echo "a=b" | sed -e s/=/:/g'],
                          python_shell=True), 'a:b')
-
-    @patch('pwd.getpwnam')
-    @patch('subprocess.Popen')
-    def test_os_environment_remains_intact(self,
-                                           popen_mock,
-                                           getpwnam_mock):
-        '''
-        Make sure the OS environment is not tainted after running a command
-        that specifies runas.
-        '''
-        environment = os.environ.copy()
-
-        popen_mock.return_value = Mock(
-            communicate=lambda *args, **kwags: ['{}', None],
-            pid=lambda: 1,
-            retcode=0
-        )
-
-        from salt.modules import cmdmod
-
-        cmdmod.__grains__ = {'os': 'Darwin', 'os_family': 'Solaris'}
-        if sys.platform.startswith(('freebsd', 'openbsd')):
-            shell = '/bin/sh'
-        else:
-            shell = '/bin/bash'
-
-        try:
-            cmdmod._run('ls',
-                        cwd=tempfile.gettempdir(),
-                        runas='foobar',
-                        shell=shell)
-
-            environment2 = os.environ.copy()
-
-            self.assertEqual(environment, environment2)
-
-            getpwnam_mock.assert_called_with('foobar')
-        finally:
-            delattr(cmdmod, '__grains__')
 
     def test_stdout(self):
         '''
@@ -276,20 +234,3 @@ class CMDModuleTest(integration.ModuleCase):
                                 f_timeout=2,
                                 python_shell=True)
         self.assertEqual(out, 'hello')
-
-    def test_run_cwd_doesnt_exist_issue_7154(self):
-        '''
-        cmd.run should fail and raise
-        salt.exceptions.CommandExecutionError if the cwd dir does not
-        exist
-        '''
-        from salt.exceptions import CommandExecutionError
-        import salt.modules.cmdmod as cmdmod
-        cmd = 'echo OHAI'
-        cwd = '/path/to/nowhere'
-        try:
-            cmdmod.run_all(cmd, cwd=cwd)
-        except CommandExecutionError:
-            pass
-        else:
-            raise RuntimeError
