+++ tests/integration/modules/test_file.py
@@ -12,11 +12,9 @@ import sys
 # Import Salt Testing libs
 import tests.integration as integration
 from tests.support.unit import skipIf
-from tests.support.mock import patch, MagicMock
 
 # Import salt libs
 import salt.utils
-from salt.modules import file as filemod
 
 
 class FileModuleTest(integration.ModuleCase):
@@ -165,59 +163,6 @@ class FileModuleTest(integration.ModuleCase):
                                                      'filehash', 'base'])
         self.assertEqual(list(ret), ['salt://http/httpd.conf', 'filehash'])
 
-    def test_source_list_for_list_returns_existing_file(self):
-        filemod.__salt__ = {
-            'cp.list_master': MagicMock(
-                return_value=['http/httpd.conf.fallback']),
-            'cp.list_master_dirs': MagicMock(return_value=[]),
-        }
-        filemod.__context__ = {}
-
-        ret = filemod.source_list(['salt://http/httpd.conf',
-                                   'salt://http/httpd.conf.fallback'],
-                                  'filehash', 'base')
-        self.assertEqual(list(ret), ['salt://http/httpd.conf.fallback', 'filehash'])
-
-    def test_source_list_for_list_returns_file_from_other_env(self):
-        def list_master(env):
-            dct = {'base': [], 'dev': ['http/httpd.conf']}
-            return dct[env]
-        filemod.__salt__ = {
-            'cp.list_master': MagicMock(side_effect=list_master),
-            'cp.list_master_dirs': MagicMock(return_value=[]),
-        }
-        filemod.__context__ = {}
-
-        ret = filemod.source_list(['salt://http/httpd.conf?saltenv=dev',
-                                   'salt://http/httpd.conf.fallback'],
-                                  'filehash', 'base')
-        self.assertEqual(list(ret), ['salt://http/httpd.conf?saltenv=dev', 'filehash'])
-
-    def test_source_list_for_list_returns_file_from_dict(self):
-        filemod.__salt__ = {
-            'cp.list_master': MagicMock(return_value=['http/httpd.conf']),
-            'cp.list_master_dirs': MagicMock(return_value=[]),
-        }
-        filemod.__context__ = {}
-
-        ret = filemod.source_list(
-            [{'salt://http/httpd.conf': ''}], 'filehash', 'base')
-        self.assertEqual(list(ret), ['salt://http/httpd.conf', 'filehash'])
-
-    @patch('salt.modules.file.os.remove')
-    def test_source_list_for_list_returns_file_from_dict_via_http(self, remove):
-        remove.return_value = None
-        filemod.__salt__ = {
-            'cp.list_master': MagicMock(return_value=[]),
-            'cp.list_master_dirs': MagicMock(return_value=[]),
-            'cp.cache_file': MagicMock(return_value='/tmp/http.conf'),
-        }
-        filemod.__context__ = {}
-
-        ret = filemod.source_list(
-            [{'http://t.est.com/http/httpd.conf': 'filehash'}], '', 'base')
-        self.assertEqual(list(ret), ['http://t.est.com/http/httpd.conf', 'filehash'])
-
     def test_source_list_for_single_local_file_slash_returns_unchanged(self):
         ret = self.run_function('file.source_list', [self.myfile,
                                                      'filehash', 'base'])
@@ -226,26 +171,4 @@ class FileModuleTest(integration.ModuleCase):
     def test_source_list_for_single_local_file_proto_returns_unchanged(self):
         ret = self.run_function('file.source_list', ['file://' + self.myfile,
                                                      'filehash', 'base'])
-        self.assertEqual(list(ret), ['file://' + self.myfile, 'filehash'])
-
-    def test_source_list_for_list_returns_existing_local_file_slash(self):
-        ret = filemod.source_list([self.myfile + '-foo',
-                                   self.myfile],
-                                  'filehash', 'base')
-        self.assertEqual(list(ret), [self.myfile, 'filehash'])
-
-    def test_source_list_for_list_returns_existing_local_file_proto(self):
-        ret = filemod.source_list(['file://' + self.myfile + '-foo',
-                                   'file://' + self.myfile],
-                                  'filehash', 'base')
-        self.assertEqual(list(ret), ['file://' + self.myfile, 'filehash'])
-
-    def test_source_list_for_list_returns_local_file_slash_from_dict(self):
-        ret = filemod.source_list(
-            [{self.myfile: ''}], 'filehash', 'base')
-        self.assertEqual(list(ret), [self.myfile, 'filehash'])
-
-    def test_source_list_for_list_returns_local_file_proto_from_dict(self):
-        ret = filemod.source_list(
-            [{'file://' + self.myfile: ''}], 'filehash', 'base')
         self.assertEqual(list(ret), ['file://' + self.myfile, 'filehash'])
