+++ tests/unit/modules/test_file.py
@@ -3,11 +3,13 @@
 # Import python libs
 from __future__ import absolute_import
 import os
+import shutil
 import tempfile
 import textwrap
 
 # Import Salt Testing libs
 from tests.support.mixins import LoaderModuleMockMixin
+from tests.support.paths import TMP
 from tests.support.unit import TestCase
 from tests.support.mock import MagicMock, patch
 
@@ -759,17 +761,88 @@ class FileBasicsTestCase(TestCase, LoaderModuleMockMixin):
 
     def setUp(self):
         self.directory = tempfile.mkdtemp()
+        self.addCleanup(shutil.rmtree, self.directory)
+        self.addCleanup(delattr, self, 'directory')
         with tempfile.NamedTemporaryFile(delete=False, mode='w+') as self.tfile:
             self.tfile.write('Hi hello! I am a file.')
             self.tfile.close()
-
-    def tearDown(self):
-        os.remove(self.tfile.name)
-        os.remove(self.directory + '/a_link')
-        os.rmdir(self.directory)
-        del self.tfile
+        self.addCleanup(os.remove, self.tfile.name)
+        self.addCleanup(delattr, self, 'tfile')
+        self.myfile = os.path.join(TMP, 'myfile')
+        with salt.utils.fopen(self.myfile, 'w+') as fp:
+            fp.write('Hello\n')
+        self.addCleanup(os.remove, self.myfile)
+        self.addCleanup(delattr, self, 'myfile')
 
     def test_symlink_already_in_desired_state(self):
         os.symlink(self.tfile.name, self.directory + '/a_link')
+        self.addCleanup(os.remove, self.directory + '/a_link')
         result = filemod.symlink(self.tfile.name, self.directory + '/a_link')
         self.assertTrue(result)
+
+    @patch('salt.modules.file.os.remove')
+    def test_source_list_for_list_returns_file_from_dict_via_http(self, remove):
+        remove.return_value = None
+        with patch.dict(filemod.__salt__, {'cp.list_master': MagicMock(return_value=[]),
+                                           'cp.list_master_dirs': MagicMock(return_value=[]),
+                                           'cp.cache_file': MagicMock(return_value='/tmp/http.conf')}):
+            ret = filemod.source_list(
+                [{'http://t.est.com/http/httpd.conf': 'filehash'}], '', 'base')
+            self.assertEqual(list(ret), ['http://t.est.com/http/httpd.conf', 'filehash'])
+
+    def test_source_list_for_list_returns_existing_file(self):
+        with patch.dict(filemod.__salt__, {'cp.list_master': MagicMock(return_value=['http/httpd.conf.fallback']),
+                                           'cp.list_master_dirs': MagicMock(return_value=[])}):
+            ret = filemod.source_list(['salt://http/httpd.conf',
+                                       'salt://http/httpd.conf.fallback'],
+                                      'filehash', 'base')
+            self.assertEqual(list(ret), ['salt://http/httpd.conf.fallback', 'filehash'])
+
+    def test_source_list_for_list_returns_file_from_other_env(self):
+        def list_master(env):
+            dct = {'base': [], 'dev': ['http/httpd.conf']}
+            return dct[env]
+
+        with patch.dict(filemod.__salt__, {'cp.list_master': MagicMock(side_effect=list_master),
+                                           'cp.list_master_dirs': MagicMock(return_value=[])}):
+            ret = filemod.source_list(['salt://http/httpd.conf?saltenv=dev',
+                                       'salt://http/httpd.conf.fallback'],
+                                      'filehash', 'base')
+            self.assertEqual(list(ret), ['salt://http/httpd.conf?saltenv=dev', 'filehash'])
+
+    def test_source_list_for_list_returns_file_from_dict(self):
+        with patch.dict(filemod.__salt__, {'cp.list_master': MagicMock(return_value=['http/httpd.conf']),
+                                           'cp.list_master_dirs': MagicMock(return_value=[])}):
+            ret = filemod.source_list(
+                [{'salt://http/httpd.conf': ''}], 'filehash', 'base')
+            self.assertEqual(list(ret), ['salt://http/httpd.conf', 'filehash'])
+
+    def test_source_list_for_list_returns_existing_local_file_slash(self):
+        with patch.dict(filemod.__salt__, {'cp.list_master': MagicMock(return_value=[]),
+                                           'cp.list_master_dirs': MagicMock(return_value=[])}):
+            ret = filemod.source_list([self.myfile + '-foo',
+                                       self.myfile],
+                                      'filehash', 'base')
+            self.assertEqual(list(ret), [self.myfile, 'filehash'])
+
+    def test_source_list_for_list_returns_existing_local_file_proto(self):
+        with patch.dict(filemod.__salt__, {'cp.list_master': MagicMock(return_value=[]),
+                                           'cp.list_master_dirs': MagicMock(return_value=[])}):
+            ret = filemod.source_list(['file://' + self.myfile + '-foo',
+                                       'file://' + self.myfile],
+                                      'filehash', 'base')
+            self.assertEqual(list(ret), ['file://' + self.myfile, 'filehash'])
+
+    def test_source_list_for_list_returns_local_file_slash_from_dict(self):
+        with patch.dict(filemod.__salt__, {'cp.list_master': MagicMock(return_value=[]),
+                                           'cp.list_master_dirs': MagicMock(return_value=[])}):
+            ret = filemod.source_list(
+                [{self.myfile: ''}], 'filehash', 'base')
+            self.assertEqual(list(ret), [self.myfile, 'filehash'])
+
+    def test_source_list_for_list_returns_local_file_proto_from_dict(self):
+        with patch.dict(filemod.__salt__, {'cp.list_master': MagicMock(return_value=[]),
+                                           'cp.list_master_dirs': MagicMock(return_value=[])}):
+            ret = filemod.source_list(
+                [{'file://' + self.myfile: ''}], 'filehash', 'base')
+            self.assertEqual(list(ret), ['file://' + self.myfile, 'filehash'])
